# frontend/contracts/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sadhana Card Tracker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/lifecycle/sahadeva-reviews": {
      "get": {
        "tags": [
          "admin-lifecycle"
        ],
        "summary": "Sahadeva Reviews",
        "operationId": "sahadeva_reviews_api_v1_admin_lifecycle_sahadeva_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_not_due",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Not Due"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SahadevaReviewItem"
                  },
                  "title": "Response Sahadeva Reviews Api V1 Admin Lifecycle Sahadeva Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/lifecycle/devotees/{user_id}/sahadeva-review": {
      "post": {
        "tags": [
          "admin-lifecycle"
        ],
        "summary": "Sahadeva Review",
        "operationId": "sahadeva_review_api_v1_admin_lifecycle_devotees__user_id__sahadeva_review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SahadevaReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SahadevaReviewResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/lifecycle/run-now": {
      "post": {
        "tags": [
          "admin-lifecycle"
        ],
        "summary": "Run Lifecycle Now",
        "description": "Admin diagnostic/manual retry. Production automation uses APScheduler.",
        "operationId": "run_lifecycle_now_api_v1_admin_lifecycle_run_now_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Lifecycle Now Api V1 Admin Lifecycle Run Now Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/corrections/devotees/{user_id}/cards/{card_date}": {
      "get": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Get Historical Card",
        "operationId": "admin_get_historical_card_api_v1_admin_corrections_devotees__user_id__cards__card_date__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "card_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Card Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCardPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Correct Historical Card",
        "operationId": "admin_correct_historical_card_api_v1_admin_corrections_devotees__user_id__cards__card_date__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "card_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Card Date"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalCardCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalCardCorrectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/corrections/devotees/{user_id}/category-history": {
      "get": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Get Category History",
        "operationId": "admin_get_category_history_api_v1_admin_corrections_devotees__user_id__category_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalCategoryHistoryPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/corrections/devotees/{user_id}/category": {
      "patch": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Correct Historical Category",
        "operationId": "admin_correct_historical_category_api_v1_admin_corrections_devotees__user_id__category_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalCategoryCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalCategoryCorrectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardStatus": {
        "type": "string",
        "enum": [
          "IN_PROGRESS",
          "FINALIZED"
        ],
        "title": "CardStatus"
      },
...
      "CategoryHistoryEntryPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "previous_category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Category Code"
          },
          "new_category_code": {
            "type": "string",
            "title": "New Category Code"
          },
          "effective_from_week": {
            "type": "string",
            "format": "date",
            "title": "Effective From Week"
          },
          "change_source": {
            "type": "string",
            "title": "Change Source"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "changed_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "previous_category_code",
          "new_category_code",
          "effective_from_week",
          "change_source",
          "reason"
        ],
        "title": "CategoryHistoryEntryPublic"
      },
...
      "DailyActivityPublic": {
        "properties": {
          "entry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entry Id"
          },
          "activity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Activity Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "scoring_type": {
            "$ref": "#/components/schemas/ScoringType"
          },
          "weekly_aggregation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AggregationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_system_derived": {
            "type": "boolean",
            "title": "Is System Derived"
          },
          "counts_toward_card_fill": {
            "type": "boolean",
            "title": "Counts Toward Card Fill"
          },
          "is_filled": {
            "type": "boolean",
            "title": "Is Filled"
          },
          "daily_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Score"
          },
          "max_score_snapshot": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Score Snapshot"
          },
          "score_calculated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Calculated At"
          },
          "score_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Details"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityValuePublic"
            },
            "type": "array",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "activity_id",
          "code",
          "name",
          "category",
          "scoring_type",
          "weekly_aggregation",
          "is_system_derived",
          "counts_toward_card_fill",
          "is_filled",
          "daily_score",
          "max_score_snapshot",
          "score_calculated_at",
          "score_details",
          "values"
        ],
        "title": "DailyActivityPublic"
      },
      "DailyActivityUpdate": {
        "properties": {
          "activity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Activity Id"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityValueUpdate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "activity_id",
          "values"
        ],
        "title": "DailyActivityUpdate"
      },
...
      "DailyCardPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "card_date": {
            "type": "string",
            "format": "date",
            "title": "Card Date"
          },
          "status": {
            "$ref": "#/components/schemas/CardStatus"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          },
          "category_code": {
            "type": "string",
            "title": "Category Code"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "first_update_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Update At"
          },
          "last_update_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Update At"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "timezone_snapshot": {
            "type": "string",
            "title": "Timezone Snapshot"
          },
          "deadline_time_snapshot": {
            "type": "string",
            "format": "time",
            "title": "Deadline Time Snapshot"
          },
          "deadline_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At Utc"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityPublic"
            },
            "type": "array",
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_date",
          "status",
          "editable",
          "category_code",
          "category_name",
          "first_update_at",
          "last_update_at",
          "finalized_at",
          "timezone_snapshot",
          "deadline_time_snapshot",
          "deadline_at_utc",
          "revision_number",
          "activities"
        ],
        "title": "DailyCardPublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalCardCorrectionRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityUpdate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "activities"
        ],
        "title": "HistoricalCardCorrectionRequest",
        "description": "Admin-only correction of raw values on an already-finalized card.\n\nThe request intentionally reuses the typed DailyActivityUpdate contract so missing\nversus intentional zero/False semantics remain identical to normal devotee updates."
      },
      "HistoricalCardCorrectionResult": {
        "properties": {
          "devotee_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Devotee User Id"
          },
          "card_date": {
            "type": "string",
            "format": "date",
            "title": "Card Date"
          },
          "card": {
            "$ref": "#/components/schemas/DailyCardPublic"
          },
          "weekly_evaluation_recalculated": {
            "type": "boolean",
            "title": "Weekly Evaluation Recalculated"
          },
          "weekly_evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyEvaluationPublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "devotee_user_id",
          "card_date",
          "card",
          "weekly_evaluation_recalculated"
        ],
        "title": "HistoricalCardCorrectionResult"
      },
      "HistoricalCategoryCorrectionRequest": {
        "properties": {
          "effective_from_week": {
            "type": "string",
            "format": "date",
            "title": "Effective From Week"
          },
          "target_category_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Target Category Code"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "effective_from_week",
          "target_category_code",
          "reason"
        ],
        "title": "HistoricalCategoryCorrectionRequest",
        "description": "Correct a category transition from one week forward until the next transition.\n\nApproved policy B: the corrected category is effective from ``effective_from_week``\nand remains effective until the next actual category transition.  If a recorded next\ntransition becomes redundant because it transitions to the same corrected category,\nthat no-op history row is removed and the correction naturally continues until the\nnext different transition."
      },
      "HistoricalCategoryCorrectionResult": {
        "properties": {
          "devotee_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Devotee User Id"
          },
          "correction_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Correction Batch Id"
          },
          "effective_from_week": {
            "type": "string",
            "format": "date",
            "title": "Effective From Week"
          },
          "effective_until_exclusive": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until Exclusive"
          },
          "previous_category_code": {
            "type": "string",
            "title": "Previous Category Code"
          },
          "corrected_category_code": {
            "type": "string",
            "title": "Corrected Category Code"
          },
          "next_transition_category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Transition Category Code"
          },
          "history_rows_removed_as_redundant": {
            "type": "integer",
            "title": "History Rows Removed As Redundant"
          },
          "affected_daily_cards": {
            "type": "integer",
            "title": "Affected Daily Cards"
          },
          "affected_weekly_evaluations": {
            "type": "integer",
            "title": "Affected Weekly Evaluations"
          },
          "current_profile_updated": {
            "type": "boolean",
            "title": "Current Profile Updated"
          },
          "current_category_code": {
            "type": "string",
            "title": "Current Category Code"
          },
          "current_academic_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Academic Year"
          }
        },
        "type": "object",
        "required": [
          "devotee_user_id",
          "correction_batch_id",
          "effective_from_week",
          "previous_category_code",
          "corrected_category_code",
          "history_rows_removed_as_redundant",
          "affected_daily_cards",
          "affected_weekly_evaluations",
          "current_profile_updated",
          "current_category_code",
          "current_academic_year"
        ],
        "title": "HistoricalCategoryCorrectionResult"
      },
      "HistoricalCategoryHistoryPublic": {
        "properties": {
          "devotee_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Devotee User Id"
          },
          "current_category_code": {
            "type": "string",
            "title": "Current Category Code"
          },
          "current_academic_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Academic Year"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/CategoryHistoryEntryPublic"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "devotee_user_id",
          "current_category_code",
          "current_academic_year",
          "history"
        ],
        "title": "HistoricalCategoryHistoryPublic"
      },
...
      "SahadevaReviewItem": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "due_week": {
            "type": "string",
            "format": "date",
            "title": "Due Week"
          },
          "review_due": {
            "type": "boolean",
            "title": "Review Due"
          },
          "pending_decision": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "CONTINUE_TO_NAKULA",
                  "DEACTIVATE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Decision"
          },
          "pending_effective_week": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Effective Week"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "full_name",
          "email",
          "due_week",
          "review_due",
          "pending_decision",
          "pending_effective_week"
        ],
        "title": "SahadevaReviewItem"
      },
      "SahadevaReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "CONTINUE_TO_NAKULA",
              "DEACTIVATE"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "reason"
        ],
        "title": "SahadevaReviewRequest"
      },
      "SahadevaReviewResult": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "decision": {
            "type": "string",
            "enum": [
              "CONTINUE_TO_NAKULA",
              "DEACTIVATE"
            ],
            "title": "Decision"
          },
          "effective_from_week": {
            "type": "string",
            "format": "date",
            "title": "Effective From Week"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "decision",
          "effective_from_week",
          "scheduled_at"
        ],
        "title": "SahadevaReviewResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}